requirements.txt
langgraph
langchain
requests
beautifulsoup4
twilio
pydantic
python-dotenv
ollama